db: drop special case for top level UDTs
This was originally done in 7f64a6ec4bda96b53dc2f7fbc60b7ad5fd900c60,
but that commit was reverted in reverted in
8517eecc2883945dc56d2fbb49d9f8244ca875bc.

The revert was done because the original change would call parse_raw
for non UDT types. Unlike the old patch, this one doesn't change the
behavior of non UDT types.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>