treewide: use mutation reader forwarding explicitly in core code
As a first step towards fixing issue 5441, which states that
mutation_reader::forwarding parameter should not have a default value
due to performance reasons, all core code (i.e. everything but tests)
usages of this parameter are now made explicit.
Default values are still preserved, because tests are a very heavy
user of this API.

Refs #5441