repair: Do not return working_row_buf_nr in get combined row hash verb
In commit b463d7039cd20445a52191bb94770d87a025e78a (repair: Introduce
get_combined_row_hash_response), working_row_buf_nr is returned in
REPAIR_GET_COMBINED_ROW_HASH in addition to the combined hash. It is
scheduled to be part of 3.1 release. However it is not backported to 3.1
by accident.

In order to be compatible between 3.1 and 3.2 repair. We need to drop
the working_row_buf_nr in 3.2 release.

Fixes: #5490
Backports: 3.1 and 3.2
Tests: Run repair in a mixed 3.1 and 3.2 cluster